implementation/requirements.txt
# DDN AI Test Failure Analysis - Python Dependencies
# Phase 2 Complete - Install with: pip install -r requirements.txt

# Core Framework
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1

# LangChain & LangGraph
langgraph==0.2.45
langchain==0.3.13
langchain-anthropic==0.3.5
langchain-openai==0.2.14
langchain-pinecone==0.2.0
langchain-community==0.3.13

# AI & ML
anthropic==0.40.0
openai==1.54.0

# Vector Database
pinecone-client==5.0.1

# Databases
pymongo==4.10.1
psycopg2-binary==2.9.10

# HTTP & Requests
requests==2.32.3
urllib3==2.2.3

# Data Processing
pydantic==2.10.4
pydantic-settings==2.6.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Testing (Optional)
pytest==8.3.4
pytest-asyncio==0.24.0

# Logging
colorlog==6.9.0

# Production (Optional)
gunicorn==23.0.0
gevent==24.11.1
